backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
supabase==2.7.0
pydantic==2.9.0
pydantic-settings==2.5.0
httpx==0.27.0
python-multipart==0.0.9
sse-starlette==2.1.3

# AI SDKs
openai==1.55.0
google-generativeai==0.8.4
anthropic==0.50.0

# Web Search
duckduckgo-search==6.0.0

# Vercel Adapter
mangum==0.18.0